Move movement files
