+++ src/main/java/nl/mtvehicles/core/movement/PacketHandler.java
@@ -3,6 +3,7 @@ package nl.mtvehicles.core.movement;
 import io.netty.channel.ChannelDuplexHandler;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.ChannelPipeline;
+import nl.mtvehicles.core.movement.versions.*;
 import org.bukkit.entity.Player;
 import java.util.NoSuchElementException;
