+++ src/main/java/nl/mtvehicles/core/movement/versions/VehicleMovement1_16.java
@@ -1,20 +1,16 @@
-package nl.mtvehicles.core.movement;
+package nl.mtvehicles.core.movement.versions;
 
-import net.minecraft.network.protocol.game.PacketPlayInSteerVehicle;
+import net.minecraft.server.v1_16_R3.PacketPlayInSteerVehicle;
+import nl.mtvehicles.core.movement.VehicleMovement;
 import org.bukkit.Location;
-import org.bukkit.craftbukkit.v1_18_R1.entity.CraftEntity;
+import org.bukkit.craftbukkit.v1_16_R3.entity.CraftEntity;
 import org.bukkit.entity.ArmorStand;
 
-public class VehicleMovement1_18 extends VehicleMovement {
+public class VehicleMovement1_16 extends VehicleMovement {
 
     @Override
     protected void teleportSeat(ArmorStand seat, Location loc){
         teleportSeat(((CraftEntity) seat).getHandle(), loc.getX(), loc.getY(), loc.getZ(), loc.getYaw(), loc.getPitch());
-    }
-
-    @Override
-    protected String getTeleportMethod(){
-        return "a";
     }
 
     @Override
